polvo/src/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usepolvo"
version = "1.1.6"
description = "A comprehensive integration package supporting various services."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp~=3.11",
    "asyncio~=3.4",
    "cryptography~=44.0",
    "gql~=3.5",
    "pydantic-settings~=2.8",
    "pydantic[email]~=2.11",
    "python-dotenv~=1.1",
    "ratelimit~=2.2",
    "requests~=2.32",
    "tenacity~=9.1",
]
license = "MIT"
[[project.authors]]
name = "Caio Pizzol"
email = "hi@usepolvo.com"

[project.optional-dependencies]
hubspot = ["hubspot-api-client~=11.1"]
gemini = ["google-genai~=1.9"]
openai = ["openai~=1.70"]
anthropic = ["anthropic~=0.49"]
google_drive = [
    "google-api-python-client~=2.166",
    "google-auth-httplib2~=0.2",
    "google-auth-oauthlib~=1.2",
]
all = [
    "hubspot-api-client~=11.1",
    "google-genai~=1.9",
    "openai~=1.70",
    "anthropic~=0.49",
    "google-api-python-client~=2.166",
    "google-auth-httplib2~=0.2",
    "google-auth-oauthlib~=1.2",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--maxfail=2 -ra"
testpaths = ["tests"]
